# ode-wrapper/src/main/java/com/trihydro/odewrapper/docs/swagger-ui-master/dist/swagger.json
...
swagger: '2.0'
info:
  description: API Endpoints to create and manage CV TIMs from WYDOT Applications.
  version: 1.0.0
  title: WYDOT Connected Vehicle REST API
  termsOfService: http://swagger.io/terms/
  contact:
    email: kperry@trihydro.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: cvodepp01:7777
...
schemes:
- http
paths:
...
  "/incident-tim":
    get:
      tags:
      - Incidents
      summary: Returns all incident active TIMs
      operationId: getIncidentTims
      produces:
      - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            "$ref": "#/definitions/GetResponse"
    post:
      tags:
      - Incidents
      summary: Add Incident TIMs
      description: ''
      operationId: addIncident
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Incident fields
        required: true
        schema:
          "$ref": "#/definitions/IncidentObj"
      responses:
        '405':
          description: Invalid input
    put:
      tags:
      - Incidents
      summary: Add and Update Chain control TIMs
      description: ''
      operationId: updateIncident
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Incident fields
        required: true
        schema:
          "$ref": "#/definitions/IncidentObj"
      responses:
        '405':
          description: Invalid input
  "/incident-tim/{incidentId}":
    get:
      tags:
      - Incidents
      summary: Find incidents by ID
      operationId: getIncidentById
      produces:
      - application/json
      parameters:
      - name: incidentId
        in: path
        description: ID of Incidenct
        required: true
        type: string
      responses:
        '200':
          description: successful operation
          schema:
            "$ref": "#/definitions/GetResponse"
        '400':
          description: Invalid ID supplied
        '404':
          description: Incident not found
    delete:
      tags:
      - Incidents
      summary: Delete purchase order by ID
      description: For valid response try integer IDs with positive integer value.         Negative
        or non-integer values will generate API errors
      operationId: deleteIncident
      produces:
      - application/json
      parameters:
      - name: incidentId
        in: path
        description: ID of the Incident that needs to be deleted
        required: true
        type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Incident not found
...
  "/create-update-rc-tim":
    post:
      tags:
      - Road Conditions
      summary: Add and Update Road Condition TIMs
      operationId: addUpdateRoadCondition
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Road Condition Fields
        required: true
        schema:
          "$ref": "#/definitions/RoadConditionObj"
      responses:
        '405':
          description: Invalid input
  "/submit-rc-ac":
    put:
      tags:
      - Road Conditions
      summary: Add and Update Chain control TIMs
      description: ''
      operationId: clearRoadConditions
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Incident fields
        required: true
        schema:
          "$ref": "#/definitions/RoadConditionObj"
      responses:
        '405':
          description: Invalid input
...
definitions:
  RoadConditionObj:
    properties:
      timRcList:
        type: array
        items:
          "$ref": "#/definitions/ChainControl"
...
  ChainControl:
    type: object
    required:
    - fromRm
    - toRm
    - route
    - advisory
    - segment
    - direction
    properties:
      fromRm:
        type: number
        example: 350
      toRm:
        type: number
        example: 360
      route:
        type: string
        example: I-80
      segment:
        type: string
        example: TC313D
      direction:
        type: string
        example: westbound
      advisory:
        type: array
        items:
          type: number
          example: 6156
...
  IncidentObj:
    properties:
      timIncidentList:
        type: array
        items:
          "$ref": "#/definitions/Incident"
  Incident:
    type: object
    required:
    - fromRm
    - toRm
    - route
    - advisory
    - incidentId
    - direction
    - highway
    - impact
    - problem
    - effect
    - action
    - pk
    - ts
    properties:
      fromRm:
        type: number
        example: 350
      toRm:
        type: number
        example: 360
      route:
        type: string
        example: I-80
      incidentId:
        type: string
        example: IN1234
      direction:
        type: string
        example: westbound
      highway:
        type: string
        example: I-80
      effect:
        type: string
        example: leftClosed
      action:
        type: string
        example: caution
      problem:
        type: string
        example: crash
      pk:
        type: string
        example: '1234'
      ts:
        type: string
        example: '2018-04-16T19:30:05.000Z'
      advisory:
        type: array
        items:
          type: number
          example: 6156
      impact:
        type: string
        enum:
        - L
        - M
        - H
        - C
        - P
  GetResponse:
    type: object
    properties:
      activeTimId:
        type: integer
        format: int32
        example: 884
      timId:
        type: integer
        format: int32
        example: 1642778
      milepostStart:
        type: integer
        format: int32
        example: 355
      milepostStop:
        type: integer
        format: int32
        example: 360
      direction:
        type: string
        example: westbound
      startDateTime:
        type: string
        example: '2018-04-16 19:30:05'
      endDateTime:
        type: string
        example: '2018-04-17 19:30:05'
      route:
        type: string
        example: '80'
      clientId:
        type: string
        example: IN1234
      satRecordId:
        type: string
        example: 01ACDA1A
      pk:
        type: string
        example: 123
...